refactor: isolate keymanager
