+++ keymanager/src/main/java/com/blockchaintp/keymanager/BaseKeyManager.java
@@ -1,4 +1,4 @@
-package com.blockchaintp.utils;
+package com.blockchaintp.keymanager;
 
 import java.util.HashMap;
 import java.util.Map;
