+++ keymanager/src/main/java/com/blockchaintp/keymanager/DirectoryKeyManager.java
@@ -9,7 +9,7 @@
  See the License for the specific language governing permissions and
  limitations under the License.
 ------------------------------------------------------------------------------*/
-package com.blockchaintp.utils;
+package com.blockchaintp.keymanager;
 
 import java.io.File;
 import java.io.FileFilter;
