+++ keymanager/src/main/java/com/blockchaintp/keymanager/InMemoryKeyManager.java
@@ -9,7 +9,7 @@
  See the License for the specific language governing permissions and
  limitations under the License.
 ------------------------------------------------------------------------------*/
-package com.blockchaintp.utils;
+package com.blockchaintp.keymanager;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
