+++ keymanager/src/main/java/com/blockchaintp/keymanager/KeyManager.java
@@ -9,7 +9,7 @@
  See the License for the specific language governing permissions and
  limitations under the License.
 ------------------------------------------------------------------------------*/
-package com.blockchaintp.utils;
+package com.blockchaintp.keymanager;
 
 import sawtooth.sdk.signing.PublicKey;
