+++ pom.xml
@@ -95,6 +95,7 @@
   <modules>
     <module>daml-on-sawtooth-protobufs</module>
     <module>timekeeper</module>
+    <module>keymanager</module>
     <module>sawtooth-daml-tp</module>
     <module>sawtooth-daml-rpc</module>
     <module>sawtooth-daml-common</module>
