+++ timekeeper/src/main/java/com/blockchaintp/sawtooth/timekeeper/TimeKeeperRunnable.java
@@ -23,7 +23,7 @@ import org.slf4j.LoggerFactory;
 import com.blockchaintp.sawtooth.timekeeper.exceptions.TimeKeeperException;
 import com.blockchaintp.sawtooth.timekeeper.protobuf.TimeKeeperUpdate;
 import com.blockchaintp.sawtooth.timekeeper.protobuf.TimeKeeperVersion;
-import com.blockchaintp.utils.KeyManager;
+import com.blockchaintp.keymanager.KeyManager;
 import com.blockchaintp.utils.SawtoothClientUtils;
 import com.google.protobuf.ByteString;
 import com.google.protobuf.InvalidProtocolBufferException;
